REmove initialized ptr

+++ tools/embedded/server-model-manager.hpp
@@ -99,8 +99,8 @@ public:
             float          allocatedSum = 0.0f;
 			float		   lp2normSum = 0.0f;
             constexpr float LAST_EPSILON = 1.0f - FLT_EPSILON;
-			int count                = sizeof(mparams.tensor_split) / sizeof(mparams.tensor_split[0]);
-			int N = count < nd ? count : nd;
+			size_t count = sizeof(mparams.tensor_split) / sizeof(mparams.tensor_split[0]);
+			size_t N = count < nd ? count : nd;
             for (int i = 0; i < N; i++) {
 				float val = fabs(*(mparams.tensor_split + i));
 				args.tensor_split[i] = val;
